Change LanesInUse type to size_t
Change LanesInUse to type size_t from int64_t. This change
is made as per changes made to upstream at -

[1] https://gerrit.openbmc.org/c/openbmc/phosphor-dbus-interfaces/+/53650

+++ host-bmc/dbus/pcie_device.hpp
@@ -36,10 +36,10 @@ class PCIeDevice : public ItemDevice
     }
 
     /** Get lanes in use */
-    int64_t lanesInUse() const override;
+    size_t lanesInUse() const override;
 
     /** Set lanes in use */
-    int64_t lanesInUse(int64_t value) override;
+    size_t lanesInUse(size_t value) override;
 
     /** Get Generation in use */
     Generations generationInUse() const override;
